+++ R/confint.R
@@ -2,6 +2,7 @@
 confint.margins <- 
 function(object, parm, level = 0.95, ...) {
     mes <- marginal_effects(object)
+    names(mes) <- gsub("^dydx_", "", names(mes))
     pnames <- colnames(mes)
     if (missing(parm)) {
         parm <- pnames
@@ -15,7 +16,7 @@ function(object, parm, level = 0.95, ...) {
     a <- c(a, 1 - a)
     fac <- qnorm(a)
     ci <- array(NA, dim = c(length(parm), 2L), dimnames = list(parm, a))
-    variances <- attributes(object)[["variances"]]
+    variances <- unlist(object[1L, grepl("Var_dydx_", names(object), fixed = TRUE), drop = TRUE])
     ses <- if (is.null(variances)) rep(NA_real_, length(cf)) else sqrt(variances)
     ci[] <- cf + ses %o% fac
     colnames(ci) <- sprintf("%0.2f%%", 100 * a)