+++ R/marginal_effects.R
@@ -17,7 +17,7 @@
 #' }
 #'
 #' Methods are also provided for the object classes \dQuote{margins} and \dQuote{marginslist} to return simplified data frames from complete \dQuote{margins} objects.
-#' @return An data.frame with dimensions equal to \code{data}, where each row is an observation and each column is the marginal effect of that variable for the data values provided by \code{data}.
+#' @return An data frame with dimensions equal to \code{data}, where each row is an observation and each column is the marginal effect of that variable for the data values provided by \code{data}.
 #' @examples
 #' require("datasets")
 #' x <- lm(mpg ~ cyl * hp + wt, data = mtcars)
@@ -31,7 +31,7 @@
 #' require('datasets')
 #' m <- margins(lm(mpg ~ hp, data = mtcars[1:10,]))
 #' marginal_effects(m)
-#' marginal_effects(m[[1]])
+#' marginal_effects(m)
 #'
 #' @seealso \code{\link{dydx}}, \code{\link{margins}}, \code{\link{build_margins}}
 #' @keywords models