+++ R/margins.R
@@ -1,11 +1,26 @@
 #' @rdname margins
+#' @name margins-package
+#' @docType package
 #' @title Marginal Effects Estimation
-#' @description \code{margins} is an S3 generic function for building a \dQuote{margins} object from a model object. Methods are currently implemented for \dQuote{lm} (and, implicitly, \dQuote{glm}) class objects.
+#' @description This package is an R port of Stata's \samp{margins} command, implemented as an S3 generic \code{margins()} for model objects, like those of class \dQuote{lm} and \dQuote{glm}. \code{margins()} is an S3 generic function for building a \dQuote{margins} object from a model object. Methods are currently implemented for \dQuote{lm} (and, implicitly, \dQuote{glm}) class objects and support is expanding. See Details, below.
+#' 
+#' The package also provides a low-level function, \code{\link{marginal_effects}}, to estimate those quantities and return a data frame of unit-specific effects and another, \code{\link{dydx}}, to provide variable-specific derivatives from models. Some of the underlying architecture for the package is provided by the low-level function \code{\link[prediction]{prediction}}, which provides a consistent data frame interface to \code{\link[stats]{predict}} for a large number of model types.
+#' @author Thomas J. Leeper
+#' @keywords package
+#' @seealso \code{\link{margins}}, \code{\link{marginal_effects}}, \code{\link[prediction]{prediction}}
+NULL
+
+#' @export
+prediction::prediction
+#' @export
+prediction::find_data
+
+#' @rdname margins
 #' @param model A model object of class \dQuote{lm}.
 #' @param data A data frame containing the data at which to evaluate the marginal effects, as in \code{\link[stats]{predict}}. This is optional, but may be required when the underlying modelling function sets \code{model = FALSE}.
-#' @param at A list of one or more named vectors, specifically values at which to calculate the marginal effects. See \code{\link{build_datalist}} for details on use.
+#' @param at A list of one or more named vectors, specifically values at which to calculate the marginal effects. These are used to modify the value of \code{data} (see \code{\link[prediction]{build_datalist}} for details on use).
 #' @param \dots Arguments passed to methods, and in turn to \code{\link{build_margins}}.
-#' @details Methods for this generic return a \dQuote{marginslist} list of one or more \dQuote{margins} objects. A \dQuote{margins} object is a data.frame consisting of the original data, predicted values and standard errors thereof, estimated marginal effects from the model \code{model}, with attributes describing various features of the marginal effects estimates.
+#' @details Methods for this generic return a \dQuote{margins} object, which is a data frame consisting of the original data, predicted values and standard errors thereof, estimated marginal effects from the model \code{model}, with attributes describing various features of the marginal effects estimates.
 #' 
 #' Some modelling functions set \code{model = FALSE} by default. For \code{margins} to work best, this should be set to \code{TRUE}. Otherwise the \code{data} argument to \code{margins} is probably required.
 #' 
@@ -19,11 +34,11 @@
 #'   \item \dQuote{merMod}, see \code{\link[lme4]{lmer}}
 #' }
 #'
-#' The \code{margins} method for objects of class \dQuote{lm} or \dQuote{glm} simply constructs a list of data.frames (using \code{\link{build_datalist}}) and calls \code{\link{build_margins}} on each. You can call \code{build_margins} directly, but it requires the explicit specification of a dataset over which to estimate the quantities of interest.
+#' The \code{margins} method for objects of class \dQuote{lm} or \dQuote{glm} simply constructs a list of data frames (using \code{\link{build_datalist}}) and calls \code{\link{build_margins}} on each. You can call \code{build_margins} directly, but it requires the explicit specification of a dataset over which to estimate the quantities of interest.
 #' 
 #' Alternatively, you can use \code{\link{marginal_effects}} to retrieve a data frame of marginal effects without constructing a \dQuote{margins} object. That can be efficient for plotting, etc., given the time-consuming nature of variance estimation.
 #' 
-#' @return An object of class \dQuote{marginslist}, composed of one or more objects of class \dQuote{margins}.
+#' @return A data frame of class \dQuote{margins}, composed of one or more row-stacked marginal effects estimates. If \code{at = NULL} (the default), then the data frame will have a number of rows equal to the original data. Otherwise, the number of rows will be a multiple thereof based upon the intersection of values specified in \code{at}.
 #' @author Thomas J. Leeper
 #' @references
 #' Greene, W.H. 2012. Econometric Analysis, 7th Ed. Boston: Pearson.
@@ -47,7 +62,7 @@
 #' ## heteroskedasticity-consistent covariance matrix
 #' summary(margins(x, vcov = vcovHC(x)))
 #' 
-#' @seealso \code{\link{marginal_effects}}, \code{\link{build_margins}}
+#' @seealso \code{\link{marginal_effects}}, \code{\link{dydx}}, \code{\link{build_margins}}, \code{\link[prediction]{prediction}}
 #' @keywords models
 #' @export
 margins <-